Pass NEXT_STATE environment variables to pre-stop script
This change updates the director to pass three `NEXT_STATE` environment
variables to the pre-stop script. The three environment variables are:

- BOSH_VM_NEXT_STATE
- BOSH_INSTANCE_NEXT_STATE
- BOSH_DEPLOYMENT_NEXT_STATE

The valid values for these environment variables are: `keep` and
`delete`. If `keep` is specified it means that the associated resource
will not be removed. If `delete` is specified it means that the
associated object is being removed.

[#164262374](https://www.pivotaltracker.com/story/show/164262374)

Co-authored-by: James Myers <jmyers@pivotal.io>
Co-authored-by: Rebecca Putinski <rputinski@pivotal.io>